# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Email CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /emails/send:
    post:
      operationId: sendEmails
      tags: [emails]
      summary: Send emails to a list of recipients
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendEmailInput"
      responses:
        "200":
          description: Email job queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendEmailResponse"
  /emails/history:
    get:
      operationId: listEmailHistory
      tags: [emails]
      summary: List email sending history
      parameters:
        - name: accountId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
            enum: [sent, failed, pending, null]
        - name: dateFrom
          in: query
          schema:
            type: ["string", "null"]
        - name: dateTo
          in: query
          schema:
            type: ["string", "null"]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Paginated email history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailHistoryPage"
  /emails/history/{id}/retry:
    post:
      operationId: retryEmail
      tags: [emails]
      summary: Retry a failed email
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Retry queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    SendEmailInput:
      type: object
      required: [accountId, templateId, recipients, subject]
      properties:
        accountId:
          type: integer
        templateId:
          type: integer
        recipients:
          type: string
          description: "Newline-separated email addresses or Name,email pairs"
        subject:
          type: string
        htmlContent:
          type: ["string", "null"]
          description: "Override template HTML"
        delaySeconds:
          type: integer
          default: 0
          description: "Delay between sends in seconds"
    SendEmailResponse:
      type: object
      required: [queued, recipients]
      properties:
        queued:
          type: integer
        recipients:
          type: integer
        message:
          type: string
    EmailHistory:
      type: object
      required: [id, accountId, accountName, recipientEmail, subject, status, createdAt]
      properties:
        id:
          type: integer
        accountId:
          type: integer
        accountName:
          type: string
        recipientEmail:
          type: string
        recipientName:
          type: ["string", "null"]
        subject:
          type: string
        status:
          type: string
          enum: [sent, failed, pending]
        errorMessage:
          type: ["string", "null"]
        sentAt:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    EmailHistoryPage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/EmailHistory"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...