# lib/api-zod/src/generated/types/emailHistory.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * Api
 * Email CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { EmailHistoryStatus } from './emailHistoryStatus';

export interface EmailHistory {
  id: number;
  accountId: number;
  accountName: string;
  recipientEmail: string;
  /** @nullable */
  recipientName?: string | null;
  subject: string;
  status: EmailHistoryStatus;
  /** @nullable */
  errorMessage?: string | null;
  /** @nullable */
  sentAt?: string | null;
  createdAt: Date;
}